+++ src/core/jasm.cpp
@@ -130,6 +130,9 @@ CLIParser::Flags SetUpCLI(char** args, int argc)
     parser.AddFlag<FlagType::Bool>("pipelines", "Use pipeline-streams instead of intermediate files.", false);
     parser.AddFlag<FlagType::String>("working-dir", "Set the working directory of the assembler.", std::filesystem::current_path().generic_string());
     parser.AddFlag<FlagType::StringList>("redirect-stdout", "Redirect stdout and stderr to given files. If only one is provided, both get redirected to it.");
+    parser.Separator();
+    parser.AddFlag<FlagType::Bool>("store-symbols", "Only affects the executable builds. Stores the symbol information inside the resulting file.", false);
+    parser.AddFlag<FlagType::Bool>("store-name", "Store assembly name in resulting file. It might come in handy.", false);
 
     parser.BindFlag("h", "help");
     parser.BindFlag("v", "version");