+++ CHANGELOG.md
@@ -7,6 +7,23 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+- Embedded version
+- Shared Library support
+
+
+## [0.1.2] - 2025-08-19
+
+### Added
+
+- Two superinstructions
+- `store-symbols` CLI flag
+
+### Fixed
+
+### Changed
+
+- Refactored `serialization.hpp`.
+
 
 ## [0.1.1] - 2025-07-17